fix bugs in image resizing

+++ app/image_processor/imageprocessor.py
@@ -80,7 +80,7 @@ class ImageProcessor():
         """load image into NxNx3 numpy array
         """
         image = Image.open(path)
-        image = image.resize(tuple(scale * dim for dim in image.size))
+        image = image.resize(tuple(int(scale * dim) for dim in image.size))
         (im_width, im_height) = image.size
         return np.array(image.getdata()).reshape((im_height, im_width, 3)).astype(np.uint8)
